Tweaked ore generation and disabled rubracium
Limited fantasy metals to Twilight Forest, Deep Dark, and Thaumic Tinkerer's bedrock dimension.  Some higher-end metals are even further limited to just the Deep Dark and bedrock due to the fact that they are so powerful, they'd be overpowered to get so early on.

Disabled Rubracium because it seems to only be used to make quicksilver, and gain experience from abstractors.  Due to quicksilver being disabled and the numerous ways to gain experience, rubracium feels redundant and unecessary.